refactor: remove size_t from codebase

+++ engine/src/common/constants.h
@@ -6,21 +6,20 @@
 
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 
 // Embedding vector dimension produced by all-MiniLM-L6-v2.
 // kVectorMemsize is the corresponding byte footprint of one vector.
-inline constexpr size_t kVectorDim     = 384;
-inline constexpr size_t kVectorMemsize = kVectorDim * sizeof(float);
+inline constexpr uint32_t kVectorDim     = 384;
+inline constexpr uint32_t kVectorMemsize = kVectorDim * sizeof(float);
 
 // Minimum cosine similarity for a cache hit.
 inline constexpr float kSimilarityThreshold = 0.85f;
 
 // Per-tier slot limits.
 // kTotalMaxSlots is the sum; used to size the unified MemoryArena.
-inline constexpr size_t kL0Capacity = 1 << 12;  // 4'096
-inline constexpr size_t kTotalSlots = 1 << 19;  // 524'288
+inline constexpr uint32_t kL0Capacity = 1 << 12;  // 4'096
+inline constexpr uint32_t kTotalSlots = 1 << 19;  // 524'288
 
 // Maximum lifetime of a PENDING node in seconds.
 // Nodes that remain PENDING beyond this deadline are treated as stale by
